+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [UNRELEASED]
+## [5.1.1] 2026-05-21
 
 ### Fixed
 
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Security
 
+- Require authentication for reading slugs.
 - Updated dependencies to address potential security vulnerabilities.
 
 ## [5.1.0] 2025-10-13